+++ include/data/LgsWarnings.h
@@ -7,10 +7,10 @@ inline string WARNING_PLACEHOLDER = "%s";
 
 struct LgsWarning {
     string msg;
-    int warningCode;
+    size_t warningCode;
 };
 
 /** Templates warnings. Should not be returned directly, but formatted and returned as a new LgsWarning */
-inline int warningCodeStart = 10000;
+inline size_t warningCodeStart = 10000;
 inline LgsWarning W10000{.msg = "Variable '%s' is not used.", .warningCode = warningCodeStart};
