+++ include/symbols/types/LgsIterable.h
@@ -10,7 +10,7 @@ class LgsIterable : public LgsType {
 public:
     LgsType* baseType;
     size_t iterLen = 0;
-    uint16_t unpackLength = 0;
+    uint8_t unpackLength = 0;
     LgsExpr* sizeExpr = nullptr;
     bool isStatic = false;
